+++ app/src/main/java/com/sopt/now/compose/data/model/FriendInfo.kt
@@ -1,0 +1,4 @@
+package com.sopt.now.compose.data.model
+
+class FriendInfo {
+}
