+++ app/src/main/java/com/sopt/now/compose/feature/auth/login/LoginScreen.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose.presentation.auth.signup
+package com.sopt.now.compose.feature.auth.login
 
 import android.widget.Toast
 import androidx.compose.foundation.layout.Column
@@ -27,38 +27,45 @@ import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
 import com.sopt.now.compose.R
 import com.sopt.now.compose.navigation.ScreenRoute
-import com.sopt.now.compose.model.User
+import com.sopt.now.compose.data.model.User
 import kotlinx.coroutines.launch
 
 @Composable
-fun SignUpScreen(
+fun LoginScreen(
     navController: NavController,
-    signUpViewModel: SignUpViewModel = viewModel(),
+    loginViewModel: LoginViewModel = viewModel()
 ) {
-
-    val signUpState by signUpViewModel.state.collectAsStateWithLifecycle()
+    val loginState by loginViewModel.state.collectAsStateWithLifecycle()
     val scope = rememberCoroutineScope()
+
     val context = LocalContext.current
 
-    LaunchedEffect(signUpState.message) {
-        signUpState.message?.let { message ->
-            if (message == "회원가입에 성공했습니다") {
+    LaunchedEffect(loginState.message) {
+        loginState.message?.let { message ->
+            if (message == "로그인에 성공했습니다") {
                 val user = User(
-                    id = signUpState.id,
-                    pw = signUpState.pw,
-                    nickname = signUpState.nickname,
-                    mbti = signUpState.mbti
+                    id = loginState.id,
+                    pw = loginState.pw,
+                    nickname = loginState.nickname,
+                    mbti = loginState.mbti
                 )
 
                 navController.currentBackStackEntry?.savedStateHandle?.set(
                     key = "User",
                     value = user
                 )
-                navController.navigate(ScreenRoute.Login.route)
+                navController.navigate(ScreenRoute.MainPage.route)
             } else {
                 Toast.makeText(context, message, Toast.LENGTH_SHORT).show()
-                signUpViewModel.clearMessage()
+                loginViewModel.clearMessage()
             }
+        }
+    }
+
+    LaunchedEffect(true) {
+        navController.previousBackStackEntry?.savedStateHandle?.run {
+            val user = get<User>("User") ?: User("", "", "", "")
+            loginViewModel.setUserInfo(user)
         }
     }
 
@@ -70,58 +77,45 @@ fun SignUpScreen(
         Spacer(modifier = Modifier.padding(20.dp))
         Text(
             modifier = Modifier.align(Alignment.CenterHorizontally),
-            text = stringResource(id = R.string.sign_up_title),
+            text = stringResource(id = R.string.login_title),
             fontSize = 25.sp
         )
         Spacer(modifier = Modifier.weight(1f))
         Text(text = stringResource(id = R.string.id))
         TextField(
-            value = signUpState.id,
+            value = loginState.id,
             onValueChange = { id ->
-                signUpViewModel.setId(id)
+                loginViewModel.setId(id)
             },
             modifier = Modifier.fillMaxWidth(),
-            placeholder = { Text(stringResource(id = R.string.sign_up_id_hint)) }
+            placeholder = { Text(stringResource(id = R.string.login_id_hint)) }
         )
         Spacer(modifier = Modifier.padding(vertical = 20.dp))
         Text(stringResource(id = R.string.pw))
         TextField(
-            value = signUpState.pw,
+            value = loginState.pw,
             onValueChange = { pw ->
-                signUpViewModel.setPassword(pw)
+                loginViewModel.setPassword(pw)
             },
             modifier = Modifier.fillMaxWidth(),
-            placeholder = { Text(stringResource(id = R.string.sign_up_pw_hint)) },
+            placeholder = { Text(stringResource(id = R.string.login_pw_hint)) },
             visualTransformation = PasswordVisualTransformation(),
-        )
-        Spacer(modifier = Modifier.padding(vertical = 20.dp))
-        Text(stringResource(id = R.string.nickname))
-        TextField(
-            value = signUpState.nickname,
-            onValueChange = { nickname ->
-                signUpViewModel.setNickname(nickname)
-            },
-            modifier = Modifier.fillMaxWidth(),
-            placeholder = { Text(stringResource(id = R.string.sign_up_nickname_hint)) }
-        )
-        Spacer(modifier = Modifier.padding(vertical = 20.dp))
-        Text(stringResource(id = R.string.mbti))
-        TextField(
-            value = signUpState.mbti,
-            onValueChange = { mbti ->
-                signUpViewModel.setMbti(mbti)
-            },
-            modifier = Modifier.fillMaxWidth(),
-            placeholder = { Text(stringResource(id = R.string.sign_up_mbti_hint)) }
         )
         Spacer(modifier = Modifier.weight(2f))
         Button(
             onClick = {
                 scope.launch {
-                    signUpViewModel.checkSignUp()
+                    loginViewModel.checkLogin()
                 }
             },
             modifier = Modifier.fillMaxWidth(),
+        ) {
+            Text(stringResource(id = R.string.btn_login))
+        }
+        Spacer(modifier = Modifier.padding(vertical = 5.dp))
+        Button(
+            onClick = { navController.navigate(ScreenRoute.SignUp.route) },
+            modifier = Modifier.fillMaxWidth()
         ) {
             Text(stringResource(id = R.string.btn_sign_up))
         }
@@ -133,5 +127,5 @@ fun SignUpScreen(
 @Composable
 private fun PreviewLogin() {
     val navController = rememberNavController()
-    SignUpScreen(navController)
+    LoginScreen(navController)
 }
