+++ app/src/main/java/com/sopt/now/compose/feature/main/home/HomeScreen.kt
@@ -1,0 +1,4 @@
+package com.sopt.now.compose.feature.main.home
+
+class HomeScreen {
+}
