+++ app/src/main/java/com/sopt/now/compose/feature/main/mypage/MyPageState.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose.presentation.main.mypage
+package com.sopt.now.compose.feature.main.mypage
 
 data class MyPageState(
     val id: String = "",
