--- app/src/main/java/com/sopt/now/compose/model/FriendInfo.kt
@@ -1,4 +1,0 @@
-package com.sopt.now.compose.model
-
-class FriendInfo {
-}
