+++ app/src/main/java/com/sopt/now/compose/navigation/BottomNavItem.kt
@@ -1,0 +1,4 @@
+package com.sopt.now.compose.navigation
+
+class BottomNavItem {
+}
