--- app/src/main/java/com/sopt/now/compose/presentation/main/home/HomeScreen.kt
@@ -1,4 +1,0 @@
-package com.sopt.now.compose.presentation.main.home
-
-class HomeScreen {
-}
