--- app/src/main/java/com/sopt/now/compose/presentation/main/search/SearchScreen.kt
@@ -1,4 +1,0 @@
-package com.sopt.now.compose.presentation.main.search
-
-class SearchScreen {
-}
